Add support for internal ADC channels
Added support for internal ADC channels ADC1_IN16 and ADC1_IN17. This
channels allow reading the temperature sensor and the internal voltage
reference.
Having the capability of measuring the internal voltage reference
(1.21V) is very important when using Vcc as ADC Vref. Being able to read
Vref before taking an ADC sample allows accurate ADC calculations.
Being able to read the temperature can be relevant for some
applications.

Also improved the checking of ADC channel availability. The current
checking was very simplistic and kind of 'lazy' because it checked only
the length of the channels array not if the requested channel was
actually available in the Solution.