Adding DisAsmPE and preliminary docs on PE file format.
This sample utility and related documentation provides a standalone tool for understanding the structure of the NETMF PE files which represent an assembly as stored in flash on a device and used by the runtime to execute code. It is intentionally a fully standalone app that does not use any of the NETMF source code. The existing nETMF code for the PE file format has a heavy dependency on boththe desktop tools used to generate the PE files and the runtime internals itself, thus making isloated comprehension more difficult. This code is not intended as a production use tool, but rather as an informational tool for understanding the PE FILE format and limitations in isolation from the rest of the system. This should, hopefully, aid in comprehension of the rest of the code in the NETMF tools and interpreter itself.